requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
Pillow>=10.0.0
pyzbar>=0.1.9
psycopg2-binary>=2.9.9
openpyxl>=3.1.0
